iterators check if the context is canceled
It's better that each iterator's NextWithContext returns early if the
given context is already canceled. Otherwise, they try to unmarshal an
item if iter.output has a response or send an unnecessary request to
the DB.

It can be the caller's responsibility to avoid calling NextWithContext
after the context gets canceled. However, it requires this kind of
context check in every loop:

```
for ctx.Err() != nil && iter.NextWithContext(ctx, &out){
    ...
}
if ctx.Err() != nil { return ctx.Err() }
```

If NextWithContext does it instead of the caller, we can avoid
such duplicated code.