improve PagingIter.LastEvaluatedKey accuracy
Fixes #186.
Previously, PagingIter.LastEvaluatedKey always returned the LastEvaluatedKey
given to us by AWS, regardless of whether we were at the end of the iterator.

This patch improves PagingIter to automatically determine LastEvaluatedKey
based on the actual last value that we have locally evaluated.

Usually we can determine this without any extra effort, but rarely we will
need to call DescribeTable to fetch the necessary key schemas.

Note that this will not work if you exclude any primary keys via a Projection.
In this case, the old behavior will be used and a warning will be printed.